Move remaining internal/utils to editable/utils
