+++ packages/components/src/internal/components/editable/models.ts
@@ -26,7 +26,7 @@ import { QueryColumn } from '../../../public/QueryColumn';
 import { QueryModel } from '../../../public/QueryModel/QueryModel';
 import { GridData } from '../../models';
 
-import { genCellKey, sortCellKeys, parseCellKey } from '../../utils';
+import { genCellKey, parseCellKey, sortCellKeys } from './utils';
 import { getQueryColumnRenderers } from '../../global';
 import { getColDateFormat, getJsonDateTimeFormatString, parseDate } from '../../util/Date';
 import { caseInsensitive, quoteValueWithDelimiters } from '../../util/utils';
@@ -479,7 +479,7 @@ export class EditorModel
         return sortCellKeys(this.selectionCells.toArray());
     }
 
-    hasRawValue(descriptor: ValueDescriptor) {
+    hasRawValue(descriptor: ValueDescriptor): boolean {
         return descriptor && descriptor.raw != null && descriptor.raw.toString().trim() !== '';
     }
 
@@ -594,5 +594,10 @@ export interface EditorModelUpdates {
 
 export interface MessageAndValue {
     message?: CellMessage;
-    valueDescriptor: ValueDescriptor
+    valueDescriptor: ValueDescriptor;
+}
+
+export interface EditableGridModels {
+    dataModels: QueryModel[];
+    editorModels: EditorModel[];
 }
