+++ packages/components/src/internal/components/editable/utils.spec.ts
@@ -1,4 +1,5 @@
 import { fromJS, List, Map } from 'immutable';
+
 import { ExtendedMap } from '../../../public/ExtendedMap';
 
 import { QueryModel } from '../../../public/QueryModel/QueryModel';
@@ -13,19 +14,19 @@ import { QueryInfo } from '../../../public/QueryInfo';
 import { QueryColumn } from '../../../public/QueryColumn';
 import { BOOLEAN_TYPE, DATE_TYPE, INTEGER_TYPE, TEXT_TYPE } from '../domainproperties/PropDescType';
 
-import { EditorMode, EditorModel, IEditableGridLoader, ValueDescriptor } from './models';
-
-import { getEditorExportData, getUpdatedDataFromGrid, initEditableGridModel } from './utils';
+import { initEditableGridModel } from './actions';
+import { EditorMode, EditorModel, EditableGridLoader, ValueDescriptor } from './models';
+import { genCellKey, getEditorExportData, getUpdatedDataFromGrid, parseCellKey, sortCellKeys } from './utils';
 
 const MODEL_ID_LOADED = 'loaded';
 
-class MockEditableGridLoader implements IEditableGridLoader {
+class MockEditableGridLoader implements EditableGridLoader {
     columns: QueryColumn[];
     id: string;
     mode = EditorMode.Insert;
     queryInfo: QueryInfo;
 
-    constructor(queryInfo: QueryInfo, props?: Partial<IEditableGridLoader>) {
+    constructor(queryInfo: QueryInfo, props?: Partial<EditableGridLoader>) {
         this.queryInfo = queryInfo;
         this.columns = props?.columns;
         this.id = props?.id ?? 'mockEditableGridLoader';
@@ -520,5 +521,22 @@ describe('getUpdatedDataFromGrid', () => {
             queryInfo
         );
         expect(updatedData).toHaveLength(0);
+    });
+
+    test('genCellKey', () => {
+        expect(genCellKey(0, 0)).toBe('0-0');
+        expect(genCellKey(1, 2)).toBe('1-2');
+    });
+
+    test('parseCellKey', () => {
+        expect(parseCellKey('0-0').colIdx).toBe(0);
+        expect(parseCellKey('0-0').rowIdx).toBe(0);
+        expect(parseCellKey('1-2').colIdx).toBe(1);
+        expect(parseCellKey('1-2').rowIdx).toBe(2);
+    });
+
+    test('getSortedCellKeys', () => {
+        expect(sortCellKeys(['0-0', '1-1', '0-1', '1-0'])).toStrictEqual(['0-0', '1-0', '0-1', '1-1']);
+        expect(sortCellKeys(['1-1', '1-15', '0-10', '1-5'])).toStrictEqual(['1-1', '1-5', '0-10', '1-15']);
     });
 });
